Use u64 for xref entry offsets

+++ src/parser_aux.rs
@@ -528,7 +528,7 @@ pub fn decode_xref_stream(mut stream: Stream) -> Result<(Xref, Dictionary)> {
                     }
                     2 => {
                         // compressed object
-                        let container = read_big_endian_integer(&mut reader, bytes2.as_mut_slice())?;
+                        let container = read_big_endian_integer(&mut reader, bytes2.as_mut_slice())? as u32;
                         let index = read_big_endian_integer(&mut reader, bytes3.as_mut_slice())? as u16;
                         xref.insert((start + j) as u32, XrefEntry::Compressed { container, index });
                     }
@@ -543,11 +543,11 @@ pub fn decode_xref_stream(mut stream: Stream) -> Result<(Xref, Dictionary)> {
     Ok((xref, dict))
 }
 
-fn read_big_endian_integer(reader: &mut Cursor<Vec<u8>>, buffer: &mut [u8]) -> Result<u32> {
+fn read_big_endian_integer(reader: &mut Cursor<Vec<u8>>, buffer: &mut [u8]) -> Result<u64> {
     reader.read_exact(buffer)?;
     let mut value = 0;
     for &mut byte in buffer {
-        value = (value << 8) + u32::from(byte);
+        value = (value << 8) + u64::from(byte);
     }
     Ok(value)
 }
